file: GameStatable module
